refactor: move uri functions to new "daft-functions" crate
