+++ daft/daft.pyi
@@ -1053,7 +1053,6 @@ class PyExpr:
     def mean(self) -> PyExpr: ...
     def min(self) -> PyExpr: ...
     def max(self) -> PyExpr: ...
-    def hash(self, seed: Any | None = None) -> PyExpr: ...
     def minhash(
         self,
         num_hashes: int,
@@ -1149,20 +1148,6 @@ class PyExpr:
     def list_slice(self, start: PyExpr, end: PyExpr) -> PyExpr: ...
     def struct_get(self, name: str) -> PyExpr: ...
     def map_get(self, key: PyExpr) -> PyExpr: ...
-    def url_download(
-        self,
-        max_connections: int,
-        raise_error_on_failure: bool,
-        multi_thread: bool,
-        config: IOConfig,
-    ) -> PyExpr: ...
-    def url_upload(
-        self,
-        folder_location: str,
-        max_connections: int,
-        multi_thread: bool,
-        io_config: IOConfig | None,
-    ): ...
     def partitioning_days(self) -> PyExpr: ...
     def partitioning_hours(self) -> PyExpr: ...
     def partitioning_months(self) -> PyExpr: ...
@@ -1187,6 +1172,21 @@ def decimal_lit(sign: bool, digits: tuple[int, ...], exp: int) -> PyExpr: ...
 def series_lit(item: PySeries) -> PyExpr: ...
 def udf(func: Callable, expressions: list[PyExpr], return_dtype: PyDataType) -> PyExpr: ...
 def resolve_expr(expr: PyExpr, schema: PySchema) -> tuple[PyExpr, PyField]: ...
+def hash(expr: PyExpr, seed: Any | None = None) -> PyExpr: ...
+def url_download(
+    expr: PyExpr,
+    max_connections: int,
+    raise_error_on_failure: bool,
+    multi_thread: bool,
+    config: IOConfig,
+) -> PyExpr: ...
+def url_upload(
+    expr: PyExpr,
+    folder_location: str,
+    max_connections: int,
+    multi_thread: bool,
+    io_config: IOConfig | None,
+) -> PyExpr: ...
 
 class PySeries:
     @staticmethod
