+++ src/daft-functions/Cargo.toml
@@ -1,0 +1,25 @@
+[dependencies]
+common-error = {path = "../common/error", default-features = false}
+common-io-config = {path = "../common/io-config", default-features = false}
+daft-core = {path = "../daft-core", default-features = false}
+daft-dsl = {path = "../daft-dsl", default-features = false}
+daft-io = {path = "../daft-io", default-features = false}
+pyo3 = {workspace = true, optional = true}
+pyo3-log = {workspace = true, optional = true}
+typetag = "0.2.16"
+serde.workspace = true
+
+[features]
+default = ["python"]
+python = [
+  "dep:pyo3",
+  "dep:pyo3-log",
+  "common-error/python",
+  "daft-core/python",
+  "common-io-config/python"
+]
+
+[package]
+name = "daft-functions"
+edition.workspace = true
+version.workspace = true
