+++ src/daft-plan/src/logical_optimization/rules/push_down_filter.rs
@@ -379,7 +379,7 @@ mod tests {
     /// Tests that we can't pushdown a filter into a ScanOperator if it has an udf-ish expression.
     #[test]
     fn filter_with_udf_not_pushed_down_into_scan() -> DaftResult<()> {
-        let pred = daft_dsl::functions::uri::download(col("a"), 1, true, true, None);
+        let pred = daft_functions::uri::download(col("a"), 1, true, true, None);
         let plan = dummy_scan_node(dummy_scan_operator(vec![
             Field::new("a", DataType::Int64),
             Field::new("b", DataType::Utf8),
