+++ src/lib.rs
@@ -72,6 +72,7 @@ pub mod pylib {
         daft_micropartition::register_modules(_py, m)?;
         daft_scan::register_modules(_py, m)?;
         daft_scheduler::register_modules(_py, m)?;
+        daft_functions::register_modules(_py, m)?;
         m.add_wrapped(wrap_pyfunction!(version))?;
         m.add_wrapped(wrap_pyfunction!(build_type))?;
         m.add_wrapped(wrap_pyfunction!(refresh_logger))?;
